Collapse timed-check into always-be-playing
